Define parameter for set/get operation
	save writes to rdb file
	set value to be of 32 bytes

+++ src/connection.h
@@ -124,7 +124,7 @@ struct connection {
     ConnectionCallbackFunc write_handler;
     ConnectionCallbackFunc read_handler;
     int fd;
-    int count;
+    unsigned long long count;
 };
 
 #define CONFIG_BINDADDR_MAX 16
